+++ src/game/behaviors/thi_top.inc.c
@@ -1,7 +1,7 @@
 // thi_top.c.inc
 
 struct SpawnParticlesInfo D_8032F134 = {
-    0, 30, MODEL_WHITE_PARTICLE_SMALL, 0, 40, 0, 20, 40, 252, 30, 20.0f, 0.0f
+    0, 30, MODEL_WHITE_PARTICLE_SMALL, 0, 40, 0, 20, 40, (s8) 252, 30, 20.0f, 0.0f
 };
 
 UNUSED u8 unused8032F134[] = { 10, 11, 12 };
